Don't raise a type error on assoc patterns
Currently when the object to match is a list it is passed to the assoc
function regardless of whether it is an alist or not. This will signal a
type error. Because checking if the object is an alist before passing it
to assoc would be costly, instead when a type error is signaled we
return nil from assoc